fix: accout/domain/exception DBError -> shared_kernel/domain/exception
