+++ src/account/domain/exception.py
@@ -1,14 +1,6 @@
 from src.shared_kernel.domain.exception import BaseHttpException
 
 
-class DBError(BaseHttpException):
-    def __init__(
-            self, code: int, message: str, log: str, err: Exception = None
-            ) -> None:
-        super().__init__(code, message, log)
-        self.error = err
-
-
 class UserError(BaseHttpException):
     def __init__(
             self, code: int, message: str, log: str, err: Exception = None
