+++ src/shared_kernel/domain/exception.py
@@ -4,3 +4,11 @@ class BaseHttpException(Exception):
         self.message = message
         self.log = log
         self.error = None
+
+
+class DBError(BaseHttpException):
+    def __init__(
+            self, code: int, message: str, log: str, err: Exception = None
+            ) -> None:
+        super().__init__(code, message, log)
+        self.error = err
